refactor demo-app
