+++ demo-app/src/components/ClusterPopup/ClusterPopup.jsx
@@ -2,8 +2,8 @@ import React, { useContext } from 'react';
 import { ThemeContext } from 'styled-components';
 import PropTypes from 'prop-types';
 import { Box, Text } from 'grommet';
-import { TextEmphasis } from './components';
-import { getStatusCounts, STATUS_MAP } from './utils';
+import { TextEmphasis } from '..';
+import { getStatusCounts, STATUS_MAP } from '../../utils';
 
 export const ClusterPopup = ({ cluster }) => {
   const theme = useContext(ThemeContext);
