+++ demo-app/src/components/GlobalHeader/index.js
@@ -1,0 +1,1 @@
+export * from './GlobalHeader.jsx';
