+++ demo-app/src/components/index.js
@@ -1,3 +1,5 @@
-export * from './network-devices';
-export * from './servers';
-export * from './shared';
+export * from './ClusterPopup';
+export * from './ContentContainer';
+export * from './GlobalHeader';
+export * from './ReverseAnchor';
+export * from './TextEmphasis';
