--- demo-app/src/components/network-devices/index.js
@@ -1,3 +1,0 @@
-export * from './NetworkMap';
-export * from './NetworkPin';
-export * from './NetworkPopup';
