--- demo-app/src/components/servers/index.js
@@ -1,4 +1,0 @@
-export * from './ServersClusterPopup';
-export * from './DeviceGroups';
-export * from './ServersMap';
-export * from './RecentActivity';
