--- demo-app/src/components/shared/index.js
@@ -1,3 +1,0 @@
-export * from './TextEmphasis';
-export * from './ContentContainer';
-export * from './ReverseAnchor';
