+++ demo-app/src/pages/custom-pins/PinKindsMap.jsx
@@ -1,17 +1,11 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { Box } from 'grommet';
 import { Grommet } from 'grommet-icons';
-import {
-  Controls,
-  Map,
-  Marker,
-  MarkerCluster,
-  Pin,
-} from 'grommet-leaflet';
+import { Controls, Map, Marker, MarkerCluster, Pin } from 'grommet-leaflet';
 import { Java } from 'grommet-icons';
-import { userLocation } from './utils';
-import data from './data/sanfrancisco.json';
-import { generic } from './themes';
+import { userLocation } from '../../utils';
+import data from '../../data/sanfrancisco.json';
+import { generic } from '../../themes';
 
 const CoffeePin = () => (
   <Box
