+++ demo-app/src/pages/index.jsx
@@ -46,7 +46,7 @@ const pages = [
     title: 'Custom pin kinds',
     description:
       'Demo of custom pin kinds such as restaurants, retail, hospitals, etc.',
-    href: '/generic-kinds',
+    href: '/custom-pins',
   },
 ];
