+++ demo-app/src/pages/network-devices/NetworkPin.jsx
