+++ demo-app/src/pages/network-devices/NetworkPopup.jsx
