+++ demo-app/src/pages/network-devices/data/network-device-data.js
